+++ app/models/check_filter.rb
@@ -106,15 +106,12 @@ class CheckFilter < ApplicationRecord
     (variable.captured_values + check_filter_values.pluck(:value)).uniq.count { |v| (v =~ /^[-+]?[0-9]*\.?[0-9]*$/).nil? } == 0
   end
 
-  # TODO: `response` will be changed to `value` in the future
   def subquery_attribute
     case variable.variable_type
-    when 'checkbox'
-      "#{subquery_scope.table_name}.value"
     when 'file'
       "#{subquery_scope.table_name}.response_file"
     else
-      "#{subquery_scope.table_name}.response"
+      "#{subquery_scope.table_name}.value"
     end
   end