+++ app/models/validation/validators/time_of_day.rb
@@ -87,7 +87,7 @@ module Validation
       end
 
       def db_key_value_pairs(response)
-        { response: parse_time_of_day_from_hash_to_s(response) }
+        { value: parse_time_of_day_from_hash_to_s(response) }
       end
     end
   end