+++ db/migrate/20170210194216_rename_response_to_value_for_sheet_variables_and_grids.rb
@@ -1,0 +1,6 @@
+class RenameResponseToValueForSheetVariablesAndGrids < ActiveRecord::Migration[5.0]
+  def change
+    rename_column :grids, :response, :value
+    rename_column :sheet_variables, :response, :value
+  end
+end